Test: Change parameter order to increase coverage.
This test case is the same than
`case_invalid_parameters_two_not_unique`: Only the last parameter is not
unique. The `parameters` parser works on all parameters except the
latest, and the latest. This constraint of parameter unicity is checked
at both places. To increase the code coverage, we must check parameter
unicity not with the last parameter.