Token: Introduce `Span` and `Token`.
I would like to thank @fflorent for the original idea and code. The goal
is to declare a `Span` structure that can be used as an input for nom
parsers. Each `Span` collects the offset, line, and column numbers for
each slices computed by parsers. Then, the `Token` structure is used to
attached a span to any “data”.

It works. We are using `bytecount::count` to count the number of
columns. And we are using `memchr::memrchr` to count the number of
lines. All counting are made relatively to the previous span, so we are
not iterating over the whole input each time. The offset is computed
based on the `nom::Offset` features.

We can compile tagua-parser as follows:

```sh
$ cargo build --release --features simd
```

or:

```sh
$ cargo build --release --features avx
```

to respectively enable SIMD or AVX optimisations. `bytecount`
performance depends on these features.