Literal: Avoid string copies.
A string is no longer copied by default. A single quoted string is
copied if and only if it contains an escaped character, because the
resulting string must not contain the escaping character, and this is
not possible to concatenate slices, so the new string is copied inside a
`Vec<u8>`. However, if no character is espaced, or if it is a nowdoc
string, there is no longer any copy.

Behind the scene, we use a clone-on-write smart pointer:
`std::borrow::Cow`.