AST: Use SmallVec for `Name`.
The `Name` enum uses small vectors instead of vectors to represent part
of the names (if `Qualified`, `RelativeQualified`, and
`FullyQualified`). Small vectors are located on the stack if the number
of items is lower than $N$, after that they are located on the heap.

The number $N = 5$ is strictly arbitrary so far.