feat(parser) Re-implement `coalesce`.
In https://github.com/php/php-langspec/issues/201, it has been pointed
out that the `coalesce-expression` was not correctly described in the
grammar. While waiting to have an Internet connection to submit a PR
that is already ready, we can re-implement it.

A new macro: `right_to_left_binary_operation!` is proposed, and is used
to re-implement the `coalese-expression` parser.